+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Adds
 
+- Adds the `"tiktok"` option to the `Icon` component.
 - Adds the `role` prop to the `Text` component.
 
 ### Updates