+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+## 1.7.1 (August 17, 2023)
+
 ### Adds
 
 - Adds the `"TikTok"` option to the `Icon` component.